requirements.txt
streamlit==0.83.0
Pillow==8.0.1
matplotlib==3.3.2
torch==1.6.0
torchvision==0.7.0
